Massive parser speed improvement
Embarrassingly, I've just realised that the Lark parser was using Earley 
when it could be using LALR, which is an order of magnitude faster. Also 
integrating the transformer into the parser for a further (smaller) 
speed improvement. With this simple change, flitter is actually suddenly 
usable for live-coding: my example programs are reloading in less than 1 
frame of lost time.