Simplify window a bit
Stop honouring the close box on the window, there is nothing useful we 
can do with this other than crash and it would seem to trigger when 
exiting fullscreen mode on the Esc key. Also disable MSAA (not sure why 
I ever had this on) - it doesn't make sense in this context. Add GL info 
logging.