Add support for transparency
Add a new `transparency` attribute to `!material`. Drawing of objects 
that have a non-zero transparency will be deferred until the end and 
then they will be depth-sorted and drawn far-to-near with an appropriate 
alpha channel value. Refraction is *not* supported, this just fades an 
object out. Transparency is tracked by lighting group, so having 
multiple lighting groups and transparent objects will probably render 
oddly. The special emission-only group had to be ditched to make this 
work, so I've switched to short-circuiting the lighting loop in the 
shader if an instance is emission-only.