Fix to make freelist actually work
OK, so I had not fully comprehended the functioning of 
`@cython.freelist()`: it can *only* manage objects of the same size as 
the base class. So moved the `Vector` attributes into `VectorLike` and 
switched the fast numbers allocation to support up to 16 doubles, which 
covers all `Matrix44` instances - this wastes some memory, but is worth 
it.