Add new `%import from` module loading
This has necessitated a bunch of changes to the file cache and partial 
evaluation. In particular, the partial evaluator will now gather up 
defined names and re-insert them as a let expression at the bottom of 
the tree to ensure that they are not simplified away and thus not 
accessible for importing.