Optimise freelists
So, interestingly, a long freelist is not a performance win. `Node`s 
tend to get allocated in large quantities and so don't benefit from a 
freelist, `Vector`s seem to benefit most from a very short freelist - 
perhaps cache locality? They are most likely to be allocated and then 
immediately free-ed in small quantities as calculations occur.