Lots of speed optimisations to physics engine
Remove added `fadd`, `fmul` and `ftruediv` methods that were previously 
added to `Vector` as these operations are now done directly in the 
engine. Add a new `Vector.copy()` method though for a fast way to grab a 
copy of a vector - now necessary because I am breaking the immutability 
rule on vectors in the physics engine for speed.