Partially evaluate known unbound names to null
Only track unbound names where we expect them to exist, which is dynamic 
attributes, everywhere else they should be either bound to a static 
value or to `None` - indicating that they are dynamic values. When 
partially-evaluating attributes, take a second pass if we're applying to 
a single literal node and have unbound names, as we can resolve them 
against the attributes of the actual node.