Capture function lvars instead of scope
Instead of pushing all of the lvars into the scope and capturing that, 
just capture the lvars themselves. Compile the body of the function with 
the lvar names in scope so that the references work. All this saves time 
and means the function body can use fast `LocalLoad`s to access 
variables from the surrounding code.