A couple more simplifier/compiler optimizations
Immediately drop the result of a top-level search combined with 
tags/attributes/appends - these nodes will all have a parent and so will 
not be appended to the root. Do a direct `StoreGlobal` of names that 
have been removed by the simplifier rather than storing them as lvars 
and then pulling them back out and storing them.