Reworking of functions that access state (or require the cache)
Make functions that depend on either state or the cache root take a 
`Context` object as their first argument. This is much more flexible and 
much quicker for imported function calls that need to change the cache 
root. However, `Context` and `StateDict` need to be pulled out of `vm` 
in order to make package importing work. Also add a test for a 
state-requiring function, which I'd never bothered doing before.