`LocalLoad` changes to how nodes are copied
Rather than copying all nodes, just copy them if the nodes are parented. 
This means that if a local is used to append nodes to a tree repeatedly, 
they are copied after the first append. Requires new parameter to 
`copynoded()`. Also, don't bother with the old logic of not copying 
parented nodes, as this no longer applies to anywhere in the codebase.