Plane slicing is not entirely reliable
Slicing with a plane doesn't seem to like complex objects, so if `!slice` is given a bunch of models it makes more sense to slice them individually and then union them than try to slice the union. Also fix crash if any of the compound operators gets a `None` model.