Compile functions into the main program
Instead of each program having its own sub-program, function bodies are compiled into the program where they are defined and the `Function` object maintains a link to this program and the address that the function starts at. Necessitated adding a new `Exit` instruction to jump to the end of the program and return early.