Push model slicing down through boolean operations
Slicing is pretty haphazard about returning a watertight model, so it's safer to do it at the leaf nodes, before boolean logic is applied, so that if something has to be wrapped in a convex hull it is the simplest model. Also fix some other small bugs in model generation.