Be much smarter about pushing slices through boolean operations
Slice all models in a union, but only slice the first in a difference or an intersection. Don't have canvas3d try to second-guess the model operations, leave it to the models themselves to sort this shit out. Also a fix to transforming a sliced model and fix cone vertices for numerical errors around the edges the same as we've done for cylinder and sphere.