Make calling functions more consistent
Log the same errors when a function call raises an exception in the VM 
or in the simplifier. Allow the simplifier to replace bad literal calls 
with a literal null (which is what the VM returns) - the call is not 
more likely to work at runtime than it was in the simplifier.