Speed up matrix maths by speeding up creation of result matrices
Make matrices same as vectors in having an `__init__()` method instead 
of a `__cinit__()` so that creating a matrix with `__new__()` does *not* 
initialise it. This saves time when we want to create a new matrix as a 
result value from some maths. Requires each of these maths functions to 
completely full the matrix and manage its storage and length. Use a new 
`_identity()` function where we genuinely want a new identity matrix.