Replace physics extra frame calculation logic
The `start_time`-based approach is fundamentally broken when there are 
large discontinuities in `time` - e.g., when switching pages. Go back to 
using `time - last_time` for the delta. Allow an additional frame only 
if the last frame was delayed (i.e., `delta > resolution`) and the 
engine has slack (`performance > 1.5`). Need to go back to tracking the 
last value of `time` in the state. `start_time` less important now, but 
retaining it anyway and extending it to reset when the `run` changes.