Dump `--autoreset` of state
It doesn't really work as desired now that other things (`!physics` and 
`!counter`s) use state in a dynamic way. Keep as `--resetonswitch` 
instead, which was the other behaviour of this flag. Dump `StateDict()`s 
special pleading for keys that are strings beginning with an underscore. 
This was only for `_counter` anyway and we almost certainly want to 
restart the beat counter if we are resetting state. Also make 
`StateDict.clear()` consistent with `.changed_keys` (it obviously 
changes the dictionary).