Deal properly with stack exhaustion or infinite recursion
Infinite recursion we track with a call depth counter and limit (fairly 
harshly) to 500 calls deep. The next recursive call will return `null` 
and will (hopefully) cause the chain to unwind. Exhausting the value or 
lnames stacks will kill the program and log an error. The engine will 
continue running, though, so the program can be amended live and 
reloaded.