New sphere mesh algorithm
Generate a sphere by subdividing an octahedron. Segments needs to be a 
multiple of 4. Generates half as many faces as old algorithm, with more 
uniform sizes – particularly at the poles. Texture mapping is a little 
worse at the poles because of this - use a large number of segments to 
get better results.