Improve physics barrier logic
If a particle crosses a barrier then rewind to that point of striking, 
reflect the velocity and then play back forwards again. This results in 
(almost) perfect reflections. The bounce example needed to be tweaked to 
deal with the new, better, reflections. Also added checks to try not to 
store broken particle state if the simulation goes out of control.