Allow simplifier to import modules
The simplifier can now parse an imported module (if the filename is 
static), simplify that module's AST and directly import static names 
from it. This allows imported inlineable functions to be inlined. Threw 
away the old module caching behaviour as I never trusted it anyway and 
it *should* be the case that module names can *always* be statically 
imported as modules aren't allowed to reference any dynamic names or 
(during import) state.