Begin adding parser tests; fix unary +/- inconsistency
Don't have a leading +/- be part of the regex for literal numbers as 
this subtly changes the logic for numbers vs other expressions: `-5**2` 
should evaluate to -25 because it is *meant* to parse as `-(5**2)`. 
However, with the current logic, the `-` is parsed as part of the number 
and so this parses as `(-5)**2`, which evaluates to 25.