Make an attempt to inline recursive functions if some arguments literal
Now attempt recursive inlining if the arguments aren't all dynamic. 
Monitor inlining depth and raise a recursion error if it's getting out 
of hand. This also requires fixing all simplifier name binding to unwind 
the bindings if an exception is raised. New tests of this functionality 
and generally tidying-up of simplifier call tests.

Fixes #69