Separate out unloading and cache-cleaning models
Throw intermediate objects (manifolds and trimeshes) as soon as a model 
falls out of immediate use. We keep the bounds and buffers so we can 
quickly render again if the model comes back into use. Also be more 
aggressive about trying to keep the `ModelCache` total objects below the 
limit.